Realizacion de Tp03-Martino-Tomas

+++ src/main/java/com/imb2025/calificaciones/service/jpa/AlumnoServicesImpl.java
@@ -14,31 +14,43 @@ public class AlumnoServicesImpl implements IAlumnoServices {
     @Autowired
     private AlumnoRepository alumnoRepository;
 
-
+    @Override
+    public Alumno update(Long id, Alumno datosActualizados) throws Exception {
+        if (alumnoRepository.existsById(id)) {
+            datosActualizados.setId(id);
+            return alumnoRepository.save(datosActualizados);
+        } else {
+            throw new Exception("Estudiante no encontrado");
+        }
+    }
 
     @Override
-    public List<Alumno> getAll() {
+    public List<Alumno> findAll() {
         return alumnoRepository.findAll();
     }
 
     @Override
-    public Alumno save(Alumno nuevoAlumno) {
-        return alumnoRepository.save(nuevoAlumno);
-    }
-
-    @Override
-    public Alumno update(Long id, Alumno datosActualizados) {
-        return alumnoRepository.save(datosActualizados);
-    }
-
-    @Override
-    public void delete(Long id) {
-        alumnoRepository.deleteById(id);
+    public boolean existsById(Long id) {
+        return alumnoRepository.existsById(id);
     }
 
     @Override
     public Alumno findById(Long id) {
-        return alumnoRepository.findById(id).orElse(null);
+        return alumnoRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Estudiante no encontrado"));
+    }
 
+    @Override
+    public Alumno create(Alumno nuevoAlumno) {
+        return alumnoRepository.save(nuevoAlumno);
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        if (alumnoRepository.existsById(id)) {
+            alumnoRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Estudiante no encontrado");
+        }
     }
 }
